✨ Refactor metered billing report, extract section builders
