+++ src/ActionsUsageAnalyser.Domain/MeteredBillingReport/Codespaces/CodespacesConsumption.cs
@@ -1,8 +1,8 @@
-namespace ActionsUsageAnalyser.Domain.MeteredBillingReport;
+namespace ActionsUsageAnalyser.Domain.MeteredBillingReport.Codespaces;
 
 public class CodespacesConsumption
 {
     public decimal AccumulatedCost { get; set; }
-    public Dictionary<string, decimal> PricePerRepository { get; set; } = new();
+    public Dictionary<string, decimal> CostPerRepository { get; set; } = new();
     public Dictionary<string, (string unit, decimal cost)> ConsumptionPerSku { get; set; } = new();
 }
