+++ src/ActionsUsageAnalyser.Domain/MeteredBillingReport/Codespaces/CodespacesEntryDataProcessor.cs
@@ -1,4 +1,4 @@
-namespace ActionsUsageAnalyser.Domain.MeteredBillingReport;
+namespace ActionsUsageAnalyser.Domain.MeteredBillingReport.Codespaces;
 
 public class CodespacesEntryDataProcessor : IReportEntryDataProcessor
 {
@@ -9,11 +9,11 @@ public class CodespacesEntryDataProcessor : IReportEntryDataProcessor
         enterprise.CodespacesConsumptionPerOwner.TryAdd(item.Owner, new CodespacesConsumption());
         var consumptionPerOwner = enterprise.CodespacesConsumptionPerOwner[item.Owner];
         var currentCost = item.Quantity * item.PricePerUnit;
-        
+
         consumptionPerOwner.AccumulatedCost += currentCost;
-        consumptionPerOwner.PricePerRepository.TryAdd(item.RepositorySlug, 0);
-        consumptionPerOwner.PricePerRepository[item.RepositorySlug] += currentCost;
-        
+        consumptionPerOwner.CostPerRepository.TryAdd(item.RepositorySlug, 0);
+        consumptionPerOwner.CostPerRepository[item.RepositorySlug] += currentCost;
+
         consumptionPerOwner.ConsumptionPerSku.TryAdd(item.SKU, (item.UnitType, 0));
         var consumptionForSkuSoFar = consumptionPerOwner.ConsumptionPerSku[item.SKU];
         consumptionPerOwner.ConsumptionPerSku[item.SKU] = (consumptionForSkuSoFar.unit, consumptionForSkuSoFar.cost + item.Quantity * item.PricePerUnit);
