+++ src/ActionsUsageAnalyser.Domain/MeteredBillingReport/Copilot/CopilotConsumption.cs
@@ -1,4 +1,4 @@
-namespace ActionsUsageAnalyser.Domain.MeteredBillingReport;
+namespace ActionsUsageAnalyser.Domain.MeteredBillingReport.Copilot;
 
 public class CopilotConsumption
 {
