+++ src/ActionsUsageAnalyser.Domain/MeteredBillingReport/Enterprise.cs
@@ -1,3 +1,9 @@
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Actions;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Codespaces;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Copilot;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Packages;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.SharedStorage;
+
 namespace ActionsUsageAnalyser.Domain.MeteredBillingReport;
 
 public class Enterprise
