+++ src/ActionsUsageAnalyser.Domain/MeteredBillingReport/EnterpriseActionsUsageConsumptionReportAnalyzer.cs
@@ -1,3 +1,8 @@
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Actions;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Codespaces;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Copilot;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.Packages;
+using ActionsUsageAnalyser.Domain.MeteredBillingReport.SharedStorage;
 using Microsoft.Extensions.Logging;
 
 namespace ActionsUsageAnalyser.Domain.MeteredBillingReport;
@@ -28,13 +33,13 @@ public class EnterpriseActionsUsageConsumptionReportAnalyzer(
             var (start, end) = (DateTime.MaxValue, DateTime.MinValue);
             var pricePerSku = new Dictionary<string, (Product product, decimal multiplier, string unit, decimal price)>();
             var enterprise = new Enterprise();
-            
+
             await foreach (var reportItem in reportReader.ReadFromSourceAsync(dataFilePath))
             {
                 if (!dataProcessors.TryGetValue(reportItem.Product, out var dataProcessor)) continue;
                 start = reportItem.Date < start ? reportItem.Date : start;
                 end = reportItem.Date > end ? reportItem.Date : end;
-                
+
                 dataProcessor.ProcessForEnterprise(enterprise, reportItem);
                 pricePerSku.TryAdd(reportItem.SKU, (reportItem.Product, reportItem.Multiplier, reportItem.UnitType, reportItem.PricePerUnit));
             }
@@ -42,28 +47,29 @@ public class EnterpriseActionsUsageConsumptionReportAnalyzer(
             using var outputWriter = outputProvider.GetOutputWriter();
 
             outputWriter.WriteTitle(2, "Metered SKUs for this enterprise");
-            
+
             outputWriter.WriteLine($"Metered data for period: **{start:yyyy-MM-dd}** to **{end:yyyy-MM-dd}**");
 
-            outputWriter.BeginTable("Product", "SKU","Unit", "Price per unit");
+            outputWriter.BeginTable("Product", "SKU", "Unit", "Price per unit");
             foreach (var sku in pricePerSku
                          .OrderBy(kv => kv.Value.product.ToString())
                          .ThenBy(kv => kv.Value.price))
             {
                 outputWriter.WriteTableRow(sku.Value.product.ToString(), sku.Key, sku.Value.unit, sku.Value.price.ToUsString());
             }
+
             outputWriter.EndTable();
 
             outputWriter.WriteLine($"Total number of organizations: {enterprise.ActionsConsumptionPerOwner.Count}");
 
-            var actionsCost = BuildActionsConsumptionSection(outputWriter, enterprise, pricePerSku);
-            var storageCost = BuildSharedStorageConsumptionSection(outputWriter, enterprise);
-            var packagesCost = BuildPackagesConsumptionSection(outputWriter, enterprise);
-            var copilotCost =  BuildCopilotConsumptionSection(outputWriter, enterprise);
-            var codespacesCost = BuildCodespacesConsumptionSection(outputWriter, enterprise);
-            
+            var actionsCost = ActionsConsumptionReportSectionBuilder.Build(outputWriter, enterprise, pricePerSku);
+            var storageCost = SharedStorageConsumptionSectionBuilder.Build(outputWriter, enterprise);
+            var packagesCost = PackagesConsumptionReportSectionBuilder.Build(outputWriter, enterprise);
+            var copilotCost = CopilotConsumptionReportSectionBuilder.Build(outputWriter, enterprise);
+            var codespacesCost = CodespacesConsumptionReportSectionBuilder.Build(outputWriter, enterprise);
+
             outputWriter.WriteTitle(1, "Summary for this enterprise");
-            
+
             outputWriter.BeginTable("Metered Cost", "Total price");
             outputWriter.WriteTableRow("Actions", actionsCost.ToUsString());
             outputWriter.WriteTableRow("Shared storage", storageCost.ToUsString());
@@ -71,155 +77,12 @@ public class EnterpriseActionsUsageConsumptionReportAnalyzer(
             outputWriter.WriteTableRow("Copilot", copilotCost.ToUsString());
             outputWriter.WriteTableRow("Codespaces", codespacesCost.ToUsString());
             outputWriter.EndTable();
-            
+
             outputWriter.WriteLine("Grand total: " + (actionsCost + packagesCost + storageCost + copilotCost).ToUsString());
         }
         catch (Exception e)
         {
             logger.LogError(e, $"Failed to process file {dataFilePath}");
         }
-    }
-
-    private decimal BuildCodespacesConsumptionSection(IOutputWriter outputWriter, Enterprise enterprise)
-    {
-        outputWriter.WriteTitle(2, "Codespaces consumption per organization");
-            
-        var totalCodespacesConsumptionForEnterprise = 0M;
-        foreach (var owner in enterprise.CodespacesConsumptionPerOwner)
-        {
-            var totalCodespacesPriceForThisOwner = owner.Value.AccumulatedCost;
-            outputWriter.WriteTitle(3, owner.Key);
-            
-            outputWriter.WriteTitle(4, "Consumption per SKU");
-            outputWriter.BeginTable("SKU", "Unit", "Total price");
-            foreach (var sku in owner.Value.ConsumptionPerSku)
-            {
-                outputWriter.WriteTableRow(sku.Key, sku.Value.unit, sku.Value.cost.ToUsString());
-            }
-            outputWriter.EndTable();
-            
-            outputWriter.WriteTitle(4, "Top 3 repositories by codespaces cost");
-            outputWriter.BeginTable("Repository", "Total price");
-            foreach (var repository in owner.Value.PricePerRepository.OrderByDescending(x => x.Value).Take(3))
-            {
-                outputWriter.WriteTableRow(repository.Key, repository.Value.ToUsString());
-            }
-            outputWriter.EndTable();
-            
-            outputWriter.WriteLine($"Total codespaces cost for this organization: {totalCodespacesPriceForThisOwner.ToUsString()}");
-            totalCodespacesConsumptionForEnterprise += totalCodespacesPriceForThisOwner;
-        }
-        outputWriter.WriteLine(string.Empty);
-        outputWriter.WriteLine($"Total codespaces consumption for the enterprise: {totalCodespacesConsumptionForEnterprise.ToUsString()}");
-        return totalCodespacesConsumptionForEnterprise;
-        
-    }
-
-    private static decimal BuildActionsConsumptionSection(IOutputWriter outputWriter, Enterprise enterprise, Dictionary<string, (Product product, decimal multiplier, string unit, decimal price)> pricePerSku)
-    {
-        outputWriter.WriteTitle(2, "Actions consumption per organization");
-
-        var totalConsumptionForEnterprise = 0M;
-        foreach (var owner in enterprise.ActionsConsumptionPerOwner)
-        {
-            var totalPriceForThisOwner = 0M;
-            outputWriter.WriteTitle(3, owner.Key);
-            outputWriter.WriteTitle(4, "Consumption per SKU");
-
-            outputWriter.BeginTable("SKU", "Minutes", "Total price");
-            foreach (var sku in owner.Value.MinutesPerSku)
-            {
-                var priceForThisSku = sku.Value * pricePerSku[sku.Key].price;
-                outputWriter.WriteTableRow(sku.Key, sku.Value.ToString("N1"), priceForThisSku.ToUsString());
-                totalPriceForThisOwner += priceForThisSku;
-            }
-            outputWriter.EndTable();
-
-            outputWriter.WriteLine($"Total cost for this organization: {totalPriceForThisOwner.ToUsString()}");
-            totalConsumptionForEnterprise += totalPriceForThisOwner;
-
-            outputWriter.WriteTitle(4, "Top 3 repositories by consumption");
-            outputWriter.BeginTable("Repository", "Total price");
-            foreach (var repository in owner.Value.PricePerRepository.OrderByDescending(x => x.Value).Take(3))
-            {
-                outputWriter.WriteTableRow(repository.Key, repository.Value.ToUsString());
-            }
-            outputWriter.EndTable();
-        }
-
-        outputWriter.WriteLine($"Total consumption for the enterprise: {totalConsumptionForEnterprise.ToUsString()}");
-        return totalConsumptionForEnterprise;
-    }
-
-    private static decimal BuildSharedStorageConsumptionSection(IOutputWriter outputWriter, Enterprise enterprise)
-    {
-        outputWriter.WriteTitle(2, "Shared storage consumption per organization");
-            
-        var totalStorageConsumptionForEnterprise = 0M;
-        foreach (var owner in enterprise.StorageConsumptionPerOwner)
-        {
-            var totalStoragePriceForThisOwner = owner.Value.AccumulatedPrice;
-            outputWriter.WriteTitle(3, owner.Key);
-            
-            outputWriter.WriteTitle(4, "Top 3 repositories by storage cost");
-            outputWriter.BeginTable("Repository", "Total price");
-            foreach (var repository in owner.Value.PricePerRepository.OrderByDescending(x => x.Value).Take(3))
-            {
-                outputWriter.WriteTableRow(repository.Key, repository.Value.ToUsString());
-            }
-            outputWriter.EndTable();
-            
-            outputWriter.WriteLine($"Total storage cost for this organization: {totalStoragePriceForThisOwner.ToUsString()}");
-            totalStorageConsumptionForEnterprise += totalStoragePriceForThisOwner;
-        }
-        outputWriter.WriteLine(string.Empty);
-        outputWriter.WriteLine($"Total storage consumption for the enterprise: {totalStorageConsumptionForEnterprise.ToUsString()}");
-        return totalStorageConsumptionForEnterprise;
-    }
-
-    private static decimal BuildCopilotConsumptionSection(IOutputWriter outputWriter, Enterprise enterprise)
-    {
-        outputWriter.WriteTitle(2, "Copilot consumption per organization");
-            
-        var totalCopilotConsumptionForEnterprise = 0M;
-        foreach (var owner in enterprise.CopilotConsumptionPerOwner)
-        {
-            var totalCopilotPriceForThisOwner = owner.Value.AccumulatedCost;
-            outputWriter.WriteTitle(3, owner.Key);
-                
-            outputWriter.WriteLine($"Total copilot cost for this organization: {totalCopilotPriceForThisOwner.ToUsString()}");
-            totalCopilotConsumptionForEnterprise += totalCopilotPriceForThisOwner;
-        }
-        outputWriter.WriteLine(string.Empty);
-        outputWriter.WriteLine($"Total copilot consumption for the enterprise: {totalCopilotConsumptionForEnterprise.ToUsString()}");
-        return totalCopilotConsumptionForEnterprise;
-    }
-
-    private static decimal BuildPackagesConsumptionSection(IOutputWriter outputWriter, Enterprise enterprise)
-    {
-        outputWriter.WriteTitle(2, "Packages consumption per organization");
-            
-        var totalPackagesConsumptionForEnterprise = 0M;
-
-        foreach (var owner in enterprise.PackagesConsumptionPerOwner)
-        {
-            var totalPackagesPriceForThisOwner = owner.Value.AccumulatedCost;
-                
-            outputWriter.WriteTitle(3, owner.Key);
-
-            outputWriter.WriteTitle(4, "Top 3 sources by packages cost");
-            outputWriter.BeginTable("Source", "Total price");
-            foreach (var repository in owner.Value.PricePerRepository.OrderByDescending(x => x.Value).Take(3))
-            {
-                outputWriter.WriteTableRow(repository.Key, repository.Value.ToUsString());
-            }
-            outputWriter.EndTable();
-                
-            outputWriter.WriteLine($"Total packages cost for this organization: {totalPackagesPriceForThisOwner.ToUsString()}");
-            totalPackagesConsumptionForEnterprise += totalPackagesPriceForThisOwner;
-        }
-        outputWriter.WriteLine(string.Empty);
-        outputWriter.WriteLine($"Total packages consumption for the enterprise: {totalPackagesConsumptionForEnterprise.ToUsString()}");
-        return totalPackagesConsumptionForEnterprise;
     }
 }
